requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
faker==12.0.1
fastapi-users-db-sqlalchemy==4.0.3
fastapi-users[sqlalchemy]==10.0.4
fastapi==0.78.0
flake8==4.0.1
freezegun==1.2.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1 
makefun==1.13.1
mako==1.2.0
markupsafe==2.1.1
mccabe==0.6.1
mixer==7.2.2
packaging==21.3; python_version >= '3.6'
passlib[bcrypt]==1.7.4
pluggy==1.0.0
asyncpg==0.27.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyflakes==2.4.0
pyjwt[crypto]==2.3.0
pyparsing==3.0.9
pytest-asyncio==0.18.3
pytest-freezegun==0.4.2
pytest-pythonpath==0.7.4
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pyyaml==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
sqlalchemy==1.4.36
starlette==0.19.1
toml==0.10.2
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn[standard]==0.17.6
watchgod==0.8.2
websockets==10.3
